+++ test/Complex.js
@@ -1,4 +1,5 @@
 /* eslint-env mocha */
+
 const algebra = require('algebra')
 
 const C = algebra.Complex