+++ test/CompositionAlgebra.js
@@ -1,4 +1,5 @@
 /* eslint-env mocha */
+
 const CompositionAlgebra = require('../src/CompositionAlgebra')
 const realField = require('../src/realField')