+++ test/VectorSpace.js
@@ -1,4 +1,5 @@
 /* eslint-env mocha */
+
 const algebra = require('algebra')
 const notDefined = require('not-defined')